+++ db/migrate/20190514142223_common_add_created_by_to_promo_codes.rb
@@ -1,0 +1,5 @@
+class CommonAddCreatedByToPromoCodes < ActiveRecord::Migration[5.0]
+  def change
+    add_reference :promo_codes, :created_by
+  end
+end